Pybulllet/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pybullet==3.2.6
sentence-transformers==2.2.2
torch==2.1.0
pillow==10.1.0
numpy==1.24.3
